cleanup: move custom tail calls in sys_exit
