--- driver/modern_bpf/programs/tail_called/events/custom_logic/drop.bpf.c
@@ -1,47 +1,0 @@
-// SPDX-License-Identifier: GPL-2.0-only OR MIT
-/*
- * Copyright (C) 2023 The Falco Authors.
- *
- * This file is dual licensed under either the MIT or GPL 2. See MIT.txt
- * or GPL2.txt for full copies of the license.
- */
-
-#include <helpers/interfaces/fixed_size_event.h>
-
-SEC("tp_btf/sys_exit")
-int BPF_PROG(t1_drop_e) {
-	struct ringbuf_struct ringbuf;
-	if(!ringbuf__reserve_space(&ringbuf, DROP_E_SIZE, PPME_DROP_E)) {
-		return 0;
-	}
-
-	ringbuf__store_event_header(&ringbuf);
-
-	/*=============================== COLLECT PARAMETERS ===========================*/
-
-	ringbuf__store_u32(&ringbuf, maps__get_sampling_ratio());
-
-	/*=============================== COLLECT PARAMETERS ===========================*/
-
-	ringbuf__submit_event(&ringbuf);
-	return 0;
-}
-
-SEC("tp_btf/sys_exit")
-int BPF_PROG(t1_drop_x) {
-	struct ringbuf_struct ringbuf;
-	if(!ringbuf__reserve_space(&ringbuf, DROP_X_SIZE, PPME_DROP_X)) {
-		return 0;
-	}
-
-	ringbuf__store_event_header(&ringbuf);
-
-	/*=============================== COLLECT PARAMETERS ===========================*/
-
-	ringbuf__store_u32(&ringbuf, maps__get_sampling_ratio());
-
-	/*=============================== COLLECT PARAMETERS ===========================*/
-
-	ringbuf__submit_event(&ringbuf);
-	return 0;
-}
