+++ driver/ppm_events_public.h
@@ -1536,9 +1536,6 @@ enum extra_syscall_codes {
 	T2_CLONE3_X,
 	T2_FORK_X,
 	T2_VFORK_X,
-	T1_DROP_E,
-	T1_DROP_X,
-	T1_HOTPLUG_E,
 	T1_OPEN_BY_HANDLE_AT_X,
 	T2_EXECVE_X,
 	T2_EXECVEAT_X,
