+++ userspace/libpman/src/events_prog_names.h
@@ -353,9 +353,6 @@ static const char* extra_event_prog_names[TAIL_EXTRA_EVENT_PROG_MAX] = {
         [T2_CLONE3_X] = "t2_clone3_x",
         [T2_FORK_X] = "t2_fork_x",
         [T2_VFORK_X] = "t2_vfork_x",
-        [T1_DROP_E] = "t1_drop_e",
-        [T1_DROP_X] = "t1_drop_x",
-        [T1_HOTPLUG_E] = "t1_hotplug_e",
         [T1_OPEN_BY_HANDLE_AT_X] = "t1_open_by_handle_at_x",
         [T2_EXECVE_X] = "t2_execve_x",
         [T2_EXECVEAT_X] = "t2_execveat_x",
